리팩토링 작업과 기능 개선 등등

+++ MySongSpace_back/src/main/java/com/hyeonmusic/MySongSpace/service/TrackService.java
@@ -48,65 +48,46 @@ public class TrackService {
             musicPath = fileService.uploadFile(trackUploadDTO.getTrackFile(), "music");
             coverPath = fileService.uploadFile(trackUploadDTO.getTrackCover(), "covers");
             Track track = Track.createTrack(trackUploadDTO, member, musicPath, coverPath);
+            trackRepository.save(track);
         } catch (Exception e) {
             // 이미 업로드된 파일이 있을 경우 삭제 시도
-            if (musicPath != null) {
-                fileService.deleteFile(musicPath.substring(1));
-            }
-            if (coverPath != null) {
-                fileService.deleteFile(coverPath.substring(1));
-            }
+            deleteIfFileUploaded(musicPath, coverPath);
             throw new FileUploadException(FILE_UPLOAD_FAILED);
         }
-    }
-
-    public List<TrackResponseDTO> getAllTracks(int page, String sortBy, List<Mood> moods, List<Genre> genres) {
-        Pageable pageable = PageRequest.of(page, 10, Sort.by(getSortDirection(sortBy)).descending());
-        Page<Track> trackPage = trackRepository.findTracksWithFilters(moods, genres, sortBy, pageable);
-        return trackPage.stream()
-                .map(track -> new TrackResponseDTO(
-                        track.getTrackId(),
-                        track.getTitle(),
-                        track.getDescription(),
-                        track.getCoverPath(),
-                        track.getMusicPath(),
-                        track.getDuration(),
-                        track.getMember().getUsername(),
-                        track.getGenres(),
-                        track.getMoods()
-                ))
-                .collect(Collectors.toList());
-    }
-
-    public String getSortDirection(String sortBy) {
-        return sortBy.equals("popular") ? null : "uploadedAt"; // 최신순과 인기순 모두 DESC로 설정
-    }
-
-    public TrackResponseDTO getTrackById(Long id) {
-        Track track = trackRepository.findById(id)
-                .orElseThrow(() -> new TrackNotFoundException(TRACK_NOT_FOUND));
-
-        return new TrackResponseDTO(
-                track.getTrackId(),
-                track.getTitle(),
-                track.getDescription(),
-                track.getCoverPath(),
-                track.getMusicPath(),
-                track.getDuration(),
-                track.getMember().getUsername(),
-                track.getGenres(),
-                track.getMoods()
-        );
     }
 
     @Transactional
     public void deleteTrack(Long id) {
         // ID로 트랙 조회
         Track track = trackRepository.findById(id)
-                .orElseThrow(() -> new TrackNotFoundException(TRACK_NOT_FOUND)); // 트랙이 존재하지 않으면 예외 발생
-        fileService.deleteFile(track.getMusicPath().substring(1)); // music/ 폴더에 있는 파일 삭제
-        fileService.deleteFile(track.getCoverPath().substring(1)); // covers/ 폴더에 있는 파일 삭제
+                .orElseThrow(() -> new TrackNotFoundException(TRACK_NOT_FOUND));
+        fileService.deleteFile(track.getMusicPath().substring(1));
+        fileService.deleteFile(track.getCoverPath().substring(1));
         trackRepository.delete(track); // 트랙 삭제
+    }
+
+    public List<TrackResponseDTO> getAllTracks(int page, String sortBy, List<Mood> moods, List<Genre> genres, String keyword) {
+        Pageable pageable = PageRequest.of(page, 10, Sort.by(getSortDirection(sortBy)).descending());
+        Page<Track> trackPage = trackRepository.findTracksWithFilters(moods, genres, sortBy, keyword, pageable);
+        return trackPage.stream()
+                .map(track -> TrackResponseDTO.toResponse(track))
+                .collect(Collectors.toList());
+    }
+
+    public TrackResponseDTO getTrackById(Long id) {
+        Track track = trackRepository.findById(id)
+                .orElseThrow(() -> new TrackNotFoundException(TRACK_NOT_FOUND));
+
+        return TrackResponseDTO.toResponse(track);
+    }
+
+    private String getSortDirection(String sortBy) {
+        return sortBy.equals("popular") ? "likeCount" : "uploadedAt"; // 최신순과 인기순 모두 DESC로 설정
+    }
+
+    private void deleteIfFileUploaded(String musicPath, String coverPath) {
+        if (musicPath != null) fileService.deleteFile(musicPath.substring(1));
+        if (coverPath != null) fileService.deleteFile(coverPath.substring(1));
     }
 
 }
